Updated the launching procedure for Windows
The documentation in the web site just mentions the websocket_address to be the IP Address through which the container is connected. This is straightforward for Linux, but in the case of Windows, the exercise runs on top of Windows Subsystem for Linux(WSL). This means the IP Address through which the container is connected is that of the vEthernet adapter(WSL).