New revamped static file version system
Instead of  renaming the files when modified, settings.py stores a version number (which currently is the date of the modification). This number will be added as the version number of all js and css files.

This comes in handy as developers can make all the changes they need and when they are done, they just simply have to go to the settings.py file and update it to the current date, solving the cache issue.

Furthermore, more variables can be added if more precision is needed